+++ docs/CHANGELOG.md
@@ -1,3 +1,1 @@
-## Changelog for v2.2.3
-
-- Forgot to declare FGB driver(W/R) as function
+## Changelog for v2.2.4